+++ libsrc4/nc4dispatch.c
@@ -12,7 +12,7 @@
 
 static NC_Dispatch NC4_dispatcher = {
 
-NC_DISPATCH_NC4,
+NC_FORMATX_NC4,
 
 NC4_create,
 NC4_open,